Add support for adjustable temperature light bulbs
These aren’t color bulbs, but they’re not just dimmable either so they
have to be treated separately.

It may be better to add support for temperature to the color type
instead, since most of those also support white temperature in
addition to color.

For now this should be ok for testing purposes